feat(queries): move prestige and ranking queries to the queries folder
