+++ models/prestige.js
@@ -1,9 +1,10 @@
-import database from 'infra/database';
+import db from 'infra/database';
+import query from 'queries/prestigeQueries';
 
-async function getByContentId(contentId, { transaction } = {}) {
+async function getByContentId(contentId, { transaction, database = db } = {}) {
   const result = await database.query(
     {
-      text: queryByContentId,
+      text: query.byContentId,
       values: [contentId],
     },
     { transaction }
@@ -24,29 +25,6 @@ async function getByContentId(contentId, { transaction } = {}) {
   };
 }
 
-const queryByContentId = `
-WITH content_events AS (
-  SELECT
-    id,
-    type
-  FROM
-    events
-  WHERE
-    metadata ->> 'id' = $1
-    OR metadata ->> 'content_id' = $1
-)
-SELECT
-  amount,
-  content_events.type
-FROM
-  balance_operations
-INNER JOIN
-  content_events
-ON balance_operations.originator_id = content_events.id
-WHERE balance_type = 'user:tabcoin'
-;
-`;
-
 async function getByUserId(
   userId,
   {
@@ -54,25 +32,35 @@ async function getByUserId(
     isRoot,
     limit = 10,
     transaction,
+    database = db,
   } = {}
 ) {
   const result = await database.query(
     {
-      text: queryByUserId,
+      text: query.byUserId,
       values: [userId, timeOffset, isRoot, limit],
     },
     { transaction }
   );
 
-  const length = result.rows?.length;
+  const mean = calcTabcoinsAverage(result.rows);
 
-  if (!length) {
-    return 0;
+  return calcPrestigeLevel(mean, isRoot);
+}
+
+function calcTabcoinsAverage(tabcoinsObjectArray) {
+  const length = tabcoinsObjectArray?.length;
+
+  if (!length || typeof length !== 'number') {
+    return 1; // TabCoins default balance
   }
 
-  const tabcoins = result.rows.reduce((acc, { tabcoins }) => acc + tabcoins, 0);
-  const mean = tabcoins / length;
+  const tabcoins = tabcoinsObjectArray.reduce((acc, { tabcoins }) => acc + tabcoins, 0);
 
+  return tabcoins / length;
+}
+
+function calcPrestigeLevel(mean, isRoot) {
   if (isRoot) {
     if (0.4 >= mean) return -1;
     if (1.1 >= mean) return 0;
@@ -90,54 +78,9 @@ async function getByUserId(
   return Math.ceil(mean + 2);
 }
 
-const queryByUserId = `
-WITH content_window AS ((
-  SELECT
-    id,
-    published_at
-  FROM contents
-  WHERE
-    owner_id = $1
-    AND status = 'published'
-    AND (CASE
-      WHEN $3 IS TRUE THEN parent_id IS NULL
-      WHEN $3 IS FALSE THEN parent_id IS NOT NULL
-      ELSE TRUE
-      END)
-  ORDER BY
-    published_at DESC
-  LIMIT $4 OFFSET $4
-)
-UNION
-  SELECT
-    id,
-    published_at
-  FROM
-    contents
-  WHERE
-    owner_id = $1
-    AND status = 'published'
-    AND published_at < $2
-    AND (CASE
-      WHEN $3 IS TRUE THEN parent_id IS NULL
-      WHEN $3 IS FALSE THEN parent_id IS NOT NULL
-      ELSE TRUE
-      END)
-  ORDER BY
-    published_at DESC
-  LIMIT $4
-)
-SELECT
-  get_current_balance('content:tabcoin', content_window.id) as tabcoins
-FROM
-  content_window
-ORDER BY
-  published_at DESC
-LIMIT $4
-;
-`;
-
 export default Object.freeze({
+  calcPrestigeLevel,
+  calcTabcoinsAverage,
   getByContentId,
   getByUserId,
 });
